backend/requirements.txt
jieba
neo4j
python-dotenv
scikit-learn
flask
flask-cors
gunicorn
